move tests back to src
