+++ package.json
@@ -79,7 +79,7 @@
       "json",
       "ts"
     ],
-    "rootDir": "test",
+    "rootDir": "src",
     "testRegex": ".spec.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
