+++ src/command/command.module.spec.ts
@@ -1,33 +1,33 @@
 import { Test } from '@nestjs/testing';
 import { Logger } from '@nestjs/common';
-import { EventStore } from '../../src/event-store/event-store';
-import { OwnerAggregate } from '../../src/core/aggregates/owner.aggregate';
-import { SensorAggregate } from '../../src/core/aggregates/sensor.aggregate';
-import { OwnerRepository } from '../../src/core/repositories/owner.repository';
-import { LocationBody } from '../../src/command/controller/model/location.body';
+import { EventStore } from '../event-store/event-store';
+import { OwnerAggregate } from '../core/aggregates/owner.aggregate';
+import { SensorAggregate } from '../core/aggregates/sensor.aggregate';
+import { OwnerRepository } from '../core/repositories/owner.repository';
+import { LocationBody } from './controller/model/location.body';
 import { CommandBus, CqrsModule, EventBus, EventPublisher } from '@nestjs/cqrs';
-import { SensorRepository } from '../../src/core/repositories/sensor.repository';
-import { EventStorePublisher } from '../../src/event-store/event-store.publisher';
-import { UpdateOwnerCommand } from '../../src/command/model/update-owner.command';
-import { DeleteOwnerCommand } from '../../src/command/model/delete-owner.command';
-import { CreateSensorCommand } from '../../src/command/model/create-sensor.command';
-import { UpdateSensorCommand } from '../../src/command/model/update-sensor.command';
-import { RegisterOwnerCommand } from '../../src/command/model/register-owner.command';
-import { UpdateOwnerCommandHandler } from '../../src/command/handler/update-owner.handler';
-import { DeleteOwnerCommandHandler } from '../../src/command/handler/delete-owner.handler';
-import { CreateSensorCommandHandler } from '../../src/command/handler/create-sensor.handler';
-import { UpdateSensorCommandHandler } from '../../src/command/handler/update-sensor.handler';
-import { DeleteSensorCommandHandler } from '../../src/command/handler/delete-sensor.handler';
-import { RegisterOwnerCommandHandler } from '../../src/command/handler/register-owner.handler';
-import { ActivateSensorCommandHandler } from '../../src/command/handler/activate-sensor.handler';
-import { DeactivateSensorCommandHandler } from '../../src/command/handler/deactivate-sensor.handler';
-import { CreateDatastreamCommandHandler } from '../../src/command/handler/create-datastream.handler';
-import { DeleteDataStreamCommandHandler } from '../../src/command/handler/delete-datastream.handler';
-import { ShareSensorOwnershipCommand } from '../../src/command/model/share-sensor-ownership.command';
-import { TransferSensorOwnershipCommand } from '../../src/command/model/transfer-sensor-ownership.command';
-import { UpdateSensorLocationCommandHandler } from '../../src/command/handler/update-sensor-location.handler';
-import { ShareSensorOwnershipCommandHandler } from '../../src/command/handler/share-sensor-ownership.handler';
-import { TransferSensorOwnershipCommandHandler } from '../../src/command/handler/transfer-sensor-ownership.handler';
+import { SensorRepository } from '../core/repositories/sensor.repository';
+import { EventStorePublisher } from '../event-store/event-store.publisher';
+import { UpdateOwnerCommand } from './model/update-owner.command';
+import { DeleteOwnerCommand } from './model/delete-owner.command';
+import { CreateSensorCommand } from './model/create-sensor.command';
+import { UpdateSensorCommand } from './model/update-sensor.command';
+import { RegisterOwnerCommand } from './model/register-owner.command';
+import { UpdateOwnerCommandHandler } from './handler/update-owner.handler';
+import { DeleteOwnerCommandHandler } from './handler/delete-owner.handler';
+import { CreateSensorCommandHandler } from './handler/create-sensor.handler';
+import { UpdateSensorCommandHandler } from './handler/update-sensor.handler';
+import { DeleteSensorCommandHandler } from './handler/delete-sensor.handler';
+import { RegisterOwnerCommandHandler } from './handler/register-owner.handler';
+import { ActivateSensorCommandHandler } from './handler/activate-sensor.handler';
+import { DeactivateSensorCommandHandler } from './handler/deactivate-sensor.handler';
+import { CreateDatastreamCommandHandler } from './handler/create-datastream.handler';
+import { DeleteDataStreamCommandHandler } from './handler/delete-datastream.handler';
+import { ShareSensorOwnershipCommand } from './model/share-sensor-ownership.command';
+import { TransferSensorOwnershipCommand } from './model/transfer-sensor-ownership.command';
+import { UpdateSensorLocationCommandHandler } from './handler/update-sensor-location.handler';
+import { ShareSensorOwnershipCommandHandler } from './handler/share-sensor-ownership.handler';
+import { TransferSensorOwnershipCommandHandler } from './handler/transfer-sensor-ownership.handler';
 
 const logger: Logger = new Logger();
